Move setup_logging from commandline
