+++ clinicadl/cmdline.py
@@ -11,6 +11,7 @@ from clinicadl.random_search.random_search_cli import cli as random_search_cli
 from clinicadl.train.resume_cli import cli as resume_cli
 from clinicadl.train.train_cli import cli as train_cli
 from clinicadl.tsvtools.cli import cli as tsvtools_cli
+from clinicadl.utils.maps_manager.logwriter import setup_logging
 
 CONTEXT_SETTINGS = dict(
     # Extend content width to avoid shortening of pipeline help.
@@ -18,59 +19,6 @@ CONTEXT_SETTINGS = dict(
     # Display help string with -h, in addition to --help.
     help_option_names=["-h", "--help"],
 )
-
-
-def setup_logging(verbosity: int = 0) -> None:
-    """
-    Setup ClinicaDL's logging facilities.
-    Args:
-        verbosity (int): The desired level of verbosity for logging.
-            (0 (default): WARNING, 1: INFO, 2: DEBUG)
-    """
-    from logging import (
-        DEBUG,
-        INFO,
-        WARNING,
-        Filter,
-        Formatter,
-        StreamHandler,
-        getLogger,
-    )
-    from sys import stderr, stdout
-
-    class StdLevelFilter(Filter):
-        def __init__(self, err=False):
-            super().__init__()
-            self.err = err
-
-        def filter(self, record):
-            if record.levelno <= INFO:
-                return not self.err
-            return self.err
-
-    # Cap max verbosity level to 2.
-    verbosity = min(verbosity, 2)
-
-    # Define the module level logger.
-    logger = getLogger("clinicadl")
-    logger.setLevel([WARNING, INFO, DEBUG][verbosity])
-
-    if logger.hasHandlers():
-        logger.handlers.clear()
-
-    stdout = StreamHandler(stdout)
-    stdout.addFilter(StdLevelFilter())
-    stderr = StreamHandler(stderr)
-    stderr.addFilter(StdLevelFilter(err=True))
-    # create formatter
-    formatter = Formatter("%(asctime)s - %(levelname)s: %(message)s", "%H:%M:%S")
-    # add formatter to ch
-    stdout.setFormatter(formatter)
-    stderr.setFormatter(formatter)
-    # add ch to logger
-    logger.addHandler(stdout)
-    logger.addHandler(stderr)
-    logger.propagate = False
 
 
 @click.group(context_settings=CONTEXT_SETTINGS)
